Putting AccessibleUIController into PageRenderer
