+++ Xamarin.Forms.Platform.iOS/Platform.cs
@@ -296,8 +296,6 @@ namespace Xamarin.Forms.Platform.iOS
 				SetRenderer(view, viewRenderer);
 
 				var nativeView = viewRenderer.NativeView;
-				if (viewRenderer is PageRenderer) // PageRenderer's NativeView is the PageContainer, but we need the parent View.
-					nativeView = nativeView.Superview;
 
 				_renderer.View.AddSubview(nativeView);
 				if (viewRenderer.ViewController != null)
